refactor(runner): split runner.rs into per-vendor modules
The 2026-05-12 audit (docs/reviews/crate-audits/convergio-runner.md
LOW finding at runner.rs:1) flagged the single 292-line file as
mixing dispatch, Claude argv, Copilot argv, PATH checks, and tests
near the 300-line cap. Split:

- runner/mod. Runner trait, SpawnContext, dispatch,rs
  assert_cli_on_path, smoke tests (~140 lines)
- runner/claude. ClaudeRunner argv shape (~75 lines)rs
- runner/copilot. CopilotRunner argv shape (~105 lines)rs

Behavior unchanged. Public surface unchanged: lib.rs still
re-exports ClaudeRunner, CopilotRunner, Runner, SpawnContext,
for_kind, for_kind_with_registry, and assert_cli_on_path from the
runner module. The full argv-shape suite in tests/runner_argv.rs
keeps passing as-is.
